requirements.txt
flask==3.0.0
tensorflow==2.15.0
pillow==10.1.0
numpy==1.26.4
gunicorn==21.2.0